services/consumer/requirements.txt
kafka-python==2.0.2
psycopg2-binary==2.9.9
jsonschema==4.23.0
python-dotenv==1.0.1
pytest
